backend+ai/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
itsdangerous
jinja2
PyJWT
passlib[bcrypt]

langchain
langchain-core
langchain-community
langchain-groq
langchain-openai
langchain-google-genai


faiss-cpu
sentence-transformers
pypdf
tqdm
pydantic

huggingface-hub

easyocr
pillow
react-markdown

